move backend functions to backend.c
