+++ ext/mri/cmodules/backend.c
@@ -1,0 +1,19 @@
+static VALUE arf_get_backend_count(VALUE self){
+   return Qnil;
+}
+
+static VALUE arf_get_available_backends(VALUE self){
+   return Qnil;
+}
+
+static VALUE arf_get_backend_id(VALUE self){
+   return Qnil;
+}
+
+static VALUE arf_get_active_backend(VALUE self){
+   return Qnil;
+}
+
+static VALUE arf_get_backend_device_id(VALUE self){
+   return Qnil;
+}
